+++ Core/RemoteRepository/Models/ChatReadModel.cs
@@ -8,5 +8,5 @@ public class ChatReadModel
     public string name { get; set; }
     public string? model { get; set; }
     public int context_size { get; set; }
-    public double temperature { get; set; }
+    public decimal temperature { get; set; }
 }
