+++ Core/RemoteRepository/Models/ChatUpdateModel.cs
@@ -7,5 +7,5 @@ public class ChatUpdateModel
     public string name { get; set; }
     public string? model { get; set; }
     public int context_size { get; set; }
-    public double temperature { get; set; }
+    public decimal temperature { get; set; }
 }
