+++ Core/RemoteRepository/Models/MessageCreateModel.cs
@@ -6,5 +6,5 @@ public class MessageCreateModel
     public string role { get; set; }
     public string content { get; set; }
     public string? model { get; set; }
-    public double temperature { get; set; }
+    public decimal temperature { get; set; }
 }
